Implementação da aplicação do cupom de desconto

+++ casadocodigo/src/main/java/com/snayder/casadocodigo/domain/dtos/request/CupomRequest.java
@@ -4,6 +4,7 @@ import com.snayder.casadocodigo.annotations.UniqueValue;
 import com.snayder.casadocodigo.domain.Cupom;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class CupomRequest {
@@ -13,13 +14,13 @@ public class CupomRequest {
 
 	@Positive(message = "O valor do desconto deve ser positivo.")
 	@Max(value = 20, message = "A porcetagem máxima de desconto é {value}.")
-	private double desconto;
+	private BigDecimal desconto;
 
 	@NotNull(message = "O campo validade é obrigatório.")
 	@Future(message = "A data de validade deve ser no futuro.")
 	private LocalDate validade;
 
-	public CupomRequest(String codigo, double desconto, LocalDate validade) {
+	public CupomRequest(String codigo, BigDecimal desconto, LocalDate validade) {
 		this.codigo = codigo;
 		this.desconto = desconto;
 		this.validade = validade;
